+++ attributes/resolv.rb
@@ -1,0 +1,3 @@
+default[:sys][:resolv][:servers] = Array.new
+default[:sys][:resolv][:domain] = String.new
+default[:sys][:resolv][:search] = String.new